# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "api-judge",
    "description": "Endpoints for encrypting and decrypting secrets. For use with the Manchester Codes Api Whisperer game",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/",
  "tags": [
    {
      "name": "Encrypt",
      "description": "Endpoints for encrypting secrets"
    },
    {
      "name": "Decrypt",
      "description": "Endpoints for decrypting secrets"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/get-keys": {
      "post": {
        "tags": [
          "Encrypt"
        ],
        "summary": "returns keys which contain encrypted secret",
        "parameters": [
          {
            "name": "PostSecretRequest",
            "in": "body",
            "description": "encrypt a secret and specify the number of keys and threshold for decryption",
            "schema": {
              "$ref": "#/definitions/PostSecretRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "secret successfully encrypted",
            "schema": {
              "$ref": "#/definitions/PostSecretResponse"
            }
          }
        },
        "produces": [
          "application/json"
        ]
      }
    },
    "/claim-victory": {
      "post":{
        "tags": [
          "Decrypt"
        ],
        "parameters": [
          {
            "name": "ClaimVictoryRequest",
            "in": "body",
            "description": "post keys for decryption and an api endpoint to POST the secret to",
            "schema": {
              "$ref": "#/definitions/ClaimVictoryRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "keys successfully decrypted",
            "schema": {
              "$ref": "#/definitions/ClaimVictoryResponse"
            }
          },
          "400": {
            "description": "bad request",
            "schema": {
              "$ref": "#/definitions/ClaimVictoryResponse"
            }
          }
        },
        "produces": [
          "application/json"
        ]
      }
    }
  },
  "definitions": {
    "PostSecretRequest": {
      "required": [
        "secret",
        "shares",
        "threshold"
      ],
      "properties": {
        "secret": {
          "type": "string",
          "description": "the secret to encrypt"
        },
        "shares": {
          "type": "integer",
          "description": "the number of keys to generate"
        },
        "threshold": {
          "type": "integer",
          "description": "the number of keys required to decrypt the secret"
        }
      }
    },
    "PostSecretResponse": {
      "required": [
        "keys"
      ],
      "properties": {
        "keys": {
          "type": "array",
          "items": {
            "type": "string",
            "description": "key required to decrypt secret"
          }

        }
      }
    },
    "ClaimVictoryRequest": {
      "required": ["keys", "successURL"],
      "properties": {
        "keys": {
          "type": "array",
          "items": {
            "type": "string",
            "description": "the keys to decrypt"
          }
        },
        "successURL": {
          "type": "string",
          "description": "the api endpoint to POST the decrypted secret to"
        }
      }
    },
    "ClaimVictoryResponse": {
      "required": ["message"],
      "properties": {
        "message": {
          "type": "string",
          "description": "instructions"
        }
      }
    }
  }
}